+++ crates/conjure-cp-core/src/utils/combinatorics.rs
@@ -22,7 +22,7 @@ impl CombinatoricsError {
 /// C(n, r) = n! / (r! * (n-r)!)
 ///
 /// Not defined for r > n.
-pub fn count_combinations(n_total: u64, n_choose: u64) -> Result<u64, CombinatoricsError> {
+pub fn count_combinations(n_total: i32, n_choose: i32) -> Result<i32, CombinatoricsError> {
     if n_choose > n_total {
         return Err(CombinatoricsError::not_defined(
             "n_choose must be <= n_total",
@@ -34,11 +34,11 @@ pub fn count_combinations(n_total: u64, n_choose: u64) -> Result<u64, Combinator
 
     // Repeatedly multiply / divide as factors get big fast;
     // return None if we overflow anyway
-    (1u64..=k).try_fold(1u64, |acc, val| {
+    (1i32..=k).try_fold(1i32, |acc, val| {
         n_total
             .checked_sub(val)
             .ok_or(CombinatoricsError::Overflow)? // n_total - val
-            .checked_add(1u64)
+            .checked_add(1i32)
             .ok_or(CombinatoricsError::Overflow)? // + 1
             .checked_mul(acc)
             .ok_or(CombinatoricsError::Overflow)? // * acc
@@ -55,7 +55,7 @@ pub fn count_combinations(n_total: u64, n_choose: u64) -> Result<u64, Combinator
 ///
 /// Not defined for r > n.
 #[allow(dead_code)]
-pub fn count_permutations(n_total: u64, n_choose: u64) -> Result<u64, CombinatoricsError> {
+pub fn count_permutations(n_total: i32, n_choose: i32) -> Result<i32, CombinatoricsError> {
     if n_choose > n_total {
         return Err(CombinatoricsError::not_defined(
             "n_choose must be <= n_total",
@@ -65,9 +65,9 @@ pub fn count_permutations(n_total: u64, n_choose: u64) -> Result<u64, Combinator
     let start = n_total
         .checked_sub(n_choose)
         .ok_or(CombinatoricsError::Overflow)?
-        .checked_add(1u64)
+        .checked_add(1i32)
         .ok_or(CombinatoricsError::Overflow)?;
-    (start..=n_total).try_fold(1u64, |acc, val| {
+    (start..=n_total).try_fold(1i32, |acc, val| {
         acc.checked_mul(val).ok_or(CombinatoricsError::Overflow)
     })
 }
